medical-iot-fleet/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
email-validator
pydantic-settings
httpx
paho-mqtt
aiosmtplib
python-dotenv
authlib
itsdangerous
starlette